requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-multipart==0.0.5
aiohttp==3.8.1
beautifulsoup4==4.9.3
azure-ai-formrecognizer==3.2.0
azure-core==1.24.0
PyMuPDF==1.19.6
python-dotenv==0.19.2
boto3==1.20.24
numpy==1.23.5
pandas==1.5.3
httpx==0.23.0
certifi==2021.10.8
tabula-py==2.7.0
openpyxl==3.1.2